requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
